Fix mvin specification in README
Mvout mentions that the 32nd bit of spad address marks if the destination is in the accumulator.  Mvin seems to have this semantics as well (as in https://github.com/ucb-bar/gemmini-rocc-tests/blob/3da2662f551556ce736bb215eca88af34ee13cb4/include/gemmini.h#L250 as well as stated in the following "Core Matmul Sequences" section.