Stall for WAWs between load instructions
Since tilelink can respond to loads out-of-order, we can't rely on the the load pipeline to handle WAW hazards for us.
This should not affect performance, since this only adds stalls if two load instructions in the buffer write to the same region of scratchpad.  This should rarely be the case in normal code.